pass schema to table. rename table comp, delete grid comp
